requirements.txt
Click
flask
fastapi
unvicorn
Flask-SQLAlchemy
gunicorn
itsdangerous
Jinja2
MarkupSafe
SQLAlchemy
Werkzeug
llama_index
torch
transformers
panda
numpy
langchain
python-dotenv
flask-login
python-multipart